(android) (Refactor) Move JSONObject creator functions to utilities.
